+++ fltk/examples/fb.rs
@@ -14,8 +14,8 @@ fn main() {
         fb.push(0x20); // blue
         fb.push(0xfa); // alpha
     }
-    unsafe {
-        frame.draw_framebuffer(&fb).unwrap();
-    }
+    
+    draw::draw_rgba(&mut frame, &fb).unwrap();
+
     app.run().unwrap();
 }
