+++ fltk/src/draw.rs
@@ -703,6 +703,79 @@ pub fn capture_window<Win: WindowExt>(win: &mut Win) -> Result<RgbImage, FltkErr
     }
 }
 
+/// Draw a framebuffer (rgba) into a widget
+pub fn draw_rgba<'a, T: WidgetBase>(wid: &'a mut T, fb: &'a [u8]) -> Result<(), FltkError> {
+    let width = wid.width() as u32;
+    let height = wid.height() as u32;
+    let mut img = crate::image::RgbImage::new(fb, width, height, 4)?;
+    wid.draw2(move |s| {
+        let x = s.x();
+        let y = s.y();
+        let w = s.width();
+        let h = s.height();
+        img.scale(w, h, false, true);
+        img.draw(x, y, w, h);
+    });
+    Ok(())
+}
+
+/// Draw a framebuffer (rgba) into a widget
+/// # Safety
+/// - The data passed should be valid and outlive the widget
+pub unsafe fn draw_rgba_nocopy<T: WidgetBase>(wid: &mut T, fb: &[u8]) {
+    let ptr = fb.as_ptr();
+    let len = fb.len();
+    let width = wid.width() as u32;
+    let height = wid.height() as u32; 
+    wid.draw2(move |s| {
+        let x = s.x();
+        let y = s.y();
+        let w = s.width();
+        let h = s.height();
+        if let Ok(mut img) = crate::image::RgbImage::from_data(std::slice::from_raw_parts(ptr, len), width, height, 4) {
+            img.scale(w, h, false, true);
+            img.draw(x, y, w, h);
+        }
+    });
+}
+
+
+/// Draw a framebuffer (rgba) into a widget
+pub fn draw_rgb<'a, T: WidgetBase>(wid: &'a mut T, fb: &'a [u8]) -> Result<(), FltkError> {
+    let width = wid.width() as u32;
+    let height = wid.height() as u32;
+    let mut img = crate::image::RgbImage::new(fb, width, height, 3)?;
+    wid.draw2(move |s| {
+        let x = s.x();
+        let y = s.y();
+        let w = s.width();
+        let h = s.height();
+        img.scale(w, h, false, true);
+        img.draw(x, y, w, h);
+    });
+    Ok(())
+}
+
+/// Draw a framebuffer (rgba) into a widget
+/// # Safety
+/// - The data passed should be valid and outlive the widget
+pub unsafe fn draw_rgb_nocopy<T: WidgetBase>(wid: &mut T, fb: &[u8]) {
+    let ptr = fb.as_ptr();
+    let len = fb.len();
+    let width = wid.width() as u32;
+    let height = wid.height() as u32; 
+    wid.draw2(move |s| {
+        let x = s.x();
+        let y = s.y();
+        let w = s.width();
+        let h = s.height();
+        if let Ok(mut img) = crate::image::RgbImage::from_data(std::slice::from_raw_parts(ptr, len), width, height, 3) {
+            img.scale(w, h, false, true);
+            img.draw(x, y, w, h);
+        }
+    });
+}
+
 /// Transforms raw data to png file
 pub fn write_to_png_file<I: ImageExt, P: AsRef<std::path::Path>>(
     image: &I,
